Refactor: GetPlayerInfo
