+++ src/RoomsEnglish.Api/PlayerContext/PlayerEndpoints.cs
@@ -1,8 +1,10 @@
 ﻿using AutoMapper;
+
 using MediatR;
+
 using RoomsEnglish.Api.Constants;
-using RoomsEnglish.Application.PlayerContext.GetPlayerInfo;
 using RoomsEnglish.Application.PlayerContext.UseCases.CreatePlayer;
+using RoomsEnglish.Application.PlayerContext.UseCases.GetPlayerInfo;
 using RoomsEnglish.Application.PlayerContext.UseCases.GetPlayers;
 using RoomsEnglish.Application.PlayerContext.ViewModels;
 
@@ -23,14 +25,14 @@ public static class PlayerEndpoints
         }).WithName("CreatePlayer")
             .WithTags(s_tags);
 
-       
-        // TODO: Configue api documentation
+        //Get Players
         endpoints.MapGet($"{EndpointPathMapping.Players}", async (IMapper mapper, IMediator bus) => {
             var command = new GetPlayersQuery();
             var result = await bus.Send(command);
             return result.Success ? Results.Ok(result) : Results.BadRequest(result);
         });
         
+        //Get Player Id
         endpoints.MapGet($"{EndpointPathMapping.Players}/{{id:guid}}", async (Guid id, IMapper mapper, IMediator bus) =>
         {
             //TODO: Validar o GUID
@@ -39,11 +41,9 @@ public static class PlayerEndpoints
             
             var command = mapper.Map<GetPlayerByIdQuery>(id);
             var result = await bus.Send(command);
-
-            // TODO: create a standardized return
-            await Task.Delay(1000);
-            return Results.Ok(result);
+            return result.Success ? Results.Ok(result) : Results.BadRequest(result);
         });
+
 
         endpoints.MapPut($"{EndpointPathMapping.Players}", async (PlayerViewModel playerViewModel, IMapper mapper, IMediator bus) =>
         {
