--- src/RoomsEnglish.Application/PlayerContext/GetPlayerInfo/GetPlayerByIdQuery.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-using RoomsEnglish.Application.PlayerContext.ViewModels;
-
-namespace RoomsEnglish.Application.PlayerContext.GetPlayerInfo;
-public class GetPlayerByIdQuery : IRequest<QueryResult<PlayerViewModel>>
-{
-    public Guid Id { get; init; }
-}
