+++ src/RoomsEnglish.Application/PlayerContext/UseCases/GetPlayerInfo/GetPlayerByIdQuery.cs
@@ -1,0 +1,8 @@
+﻿using RoomsEnglish.Application.PlayerContext.ViewModels;
+using RoomsEnglish.Application.SharedContext.UseCases;
+
+namespace RoomsEnglish.Application.PlayerContext.UseCases.GetPlayerInfo;
+public class GetPlayerByIdQuery : RequestQuery<PlayerViewModel>
+{
+    public Guid Id { get; init; }
+}
